feat(easy_run): adds a "run in background" option
Canvas assignments aren't always static. So, having a "run in
background" option makes it so that we can make sure to stay on
top of the assignments without having to sync it manually.

This would be ideal in a server setting, or in an idle computer
like the one I have at home.